refactor(smoothquant) compact codes
